Update on "[FX] NamedTuple repr approach 2: FakeNamedTuple type in create_arg"
Differential Revision: [D25599445](https://our.internmc.facebook.com/intern/diff/D25599445)

[ghstack-poisoned]
